requirements.txt
psycopg2==2.9.10
python-dotenv==1.1.1